chore(frontend): TTSRH-1 PR-14 — pre-push review fixups
Apply pre-push-reviewer findings (🟠×3 🟡×3 🔵×1):

- ColumnConfigurator: reorder off-by-one — после splice(-1) индексы выше
  srcIdx сдвигаются на 1 влево. Было `insertAt -= 0` (no-op), правильно
  `if (insertAt > srcIdx) insertAt -= 1`. Раньше drag с index=1 в slot=4
  приземлялся на index=5.
- ResultsTable: `sorter: { compare: () => 0 }` вместо `sorter: true` —
  suppresses Ant Table client-side sort flicker до того как server-side JQL
  ORDER BY пере-fetch'нется.
- ResultsTable.parseOrderBy: missing direction keyword (`ORDER BY summary`
  без ASC/DESC) теперь возвращает `dir: null` вместо hard-coded 'descend'.
  Header indicator не лжёт — 3-way toggle работает корректно.
- ResultsTable.SORTABLE: добавлены `'release'` и `'project'` — backend
  schema.ts declares `sortable: true` для них. Раньше клик по этим
  заголовкам silently ничего не делал.
- SearchPage: selectedRowIds clear'ятся при onJqlChange/onPageChange +
  `selectedIds` prop прокидывается в ResultsTable → controlled
  `selectedRowKeys` в rowSelection. Раньше stale IDs оставались после
  сортировки/пагинации.
- utils/saveBlob.ts: extracted helper. ExportMenu + BulkActionsBar теперь
  импортируют, раньше было 2 duplicate copy — drift risk.
- SearchPage preview-pane: placeholder "PR-14" → обновлён (drawer-preview
  не в scope §13.6, Phase 2).

См. docs/tz/TTSRH-1.md §13.6 PR-14.
